DynamicLinking: Extend "dylink" section with [] of needed libraries
This comes from my work to teach Emscripten dynamic linker to support
library -> library linkage:

	https://github.com/kripken/emscripten/pull/7512

The motivation is that without support for DSO -> DSO linking, it becomes
problematic in cases when there are several shared libraries that all
need to use another should-be shared functionality, while linking that
should-be shared functionality to main module is not an option for size
reasons. See above pull-request for more details.

In order to support library -> library linkages, we have to store
somewhere the information about which other libraries, a library needs.

This patch extends "dylink" section with such information, which is
similar to how ELF handles the same situation with DT_NEEDED entries.

/cc @sbc100