Target features section
Adds a new "target_features" custom section that lists the features an
object was built with. It is an error to link objects with mismatched
target feature sets.

The presence of the "atomics" feature also replaces the value of the
shared bit on object file memory import as the signal to the linker
that it should use shared memory. The target features section is more
generally useful and less hacky than the old solution.