Target features section (#99)
Adds a new "target_features" custom section that lists the features an
object was built with and rules controlling which features are allowed
in the final link.

The presence of the "atomics" feature also replaces the value of the
shared bit on object file memory import as the signal to the linker
that it should use shared memory. The target features section is more
generally useful and less hacky than the old solution.