Allocate magic constants for reference-types proposal
The reference types proposal adds support for multiple table indices,
requiring us to reason about tables in linking. To make sure the various
tools use consistent names/constants for the required new relocation
type, add it to the list here, giving it an authoritative number.

This is based on my WIP support for LLVM here:
https://reviews.llvm.org/D66035, but I figured it'd be good to get
this out of the way first, so we can simultaneously start adding
support to other tools and make sure everybody has consistent numbers.