Add a text format for `producers` and `dylink.0` custom sections
This proposes a text format based on the annotations proposal for two
custom sections defined in this repository. The annotations proposal is
likely to reach phase 4 and "basically done" status in the near future.
The [wasm-tools repository][tools] has implemented the text format
documented here for a bit now and it seemed best to document this
upstream as well to both gather consensus and help guide other
implementations.

[tools]: https://github.com/bytecodealliance/wasm-tools